fix vitest
