+++ template/frontend/nuxt.config.ts.jinja
@@ -2,6 +2,9 @@
 import { defineNuxtConfig } from "nuxt/config";
 export default defineNuxtConfig({
   compatibilityDate: "2024-11-01",
+  future: {
+    compatibilityVersion: 4,
+  },
   devtools: { enabled: true },
   // the conditional modules added in by the template make it complicated to format consistently...at least with only 3 'always included' modules
   // prettier-ignore
