+++ template/frontend/package.json.jinja
@@ -12,17 +12,16 @@
     "generate": "pnpm run type-check && nuxt generate",
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare && pnpm exec playwright-core install --only-shell chromium-headless-shell",
-    "test-unit": "vitest --dir tests/unit --run --coverage",
-    "test-unit:watch": "vitest --dir tests/unit",
-    "test-compiled": "vitest --test-timeout=15000 --dir tests/compiled --run",
-    "test-e2e": "{% endraw %}{% if not deploy_as_executable %}{% raw %}docker compose --file=../docker-compose.yaml build && dotenv -v USE_DOCKER_COMPOSE_FOR_VITEST_E2E=1{% endraw %}{% else %}{% raw %}dotenv -v USE_BUILT_BACKEND_FOR_VITEST_E2E=1{% endraw %}{% endif %}{% raw %} -- vitest --test-timeout=15000 --dir tests/e2e --run",
-    "test-compiled:watch": "vitest --dir tests/compiled"{% endraw %}{% if frontend_uses_graphql %}{% raw %},
+    "test-unit": "vitest --exclude=\"tests/e2e/**\" --exclude=\"tests/compiled/**\" --run --coverage",
+    "test-unit:watch": "vitest --exclude=\"tests/e2e/**\" --exclude=\"tests/compiled/**\"",
+    "test-compiled": "vitest --exclude=\"tests/e2e/**\" --exclude=\"tests/unit/**\" --test-timeout=15000 --run",
+    "test-e2e": "docker compose --file=../docker-compose.yaml build && dotenv -v USE_DOCKER_COMPOSE_FOR_VITEST_E2E=1 -- vitest --exclude=\"tests/unit/**\" --exclude=\"tests/compiled/**\" --test-timeout=15000 --run",
+    "test-compiled:watch": "vitest --exclude=\"tests/e2e/**\" --exclude=\"tests/unit/**\" --test-timeout=15000",{% endraw %}{% if frontend_uses_graphql %}{% raw %},
     "codegen": "graphql-codegen --config codegen.ts"{% endraw %}{% endif %}{% raw %}
   },
   "dependencies": {
     "@nuxt/ui": "{% endraw %}{{ nuxt_ui_version }}{% raw %}",
     "nuxt": "{% endraw %}{{ nuxt_version }}{% raw %}",
-    "typescript": "{% endraw %}{{ typescript_version }}{% raw %}",
     "vue": "{% endraw %}{{ vue_version }}{% raw %}",
     "vue-router": "{% endraw %}{{ vue_router_version }}{% raw %}"
   },
@@ -39,6 +38,7 @@
     "@nuxt/test-utils": "^3.17.2",{% endraw %}{% if frontend_uses_graphql %}{% raw %}
     "@nuxtjs/apollo": "5.0.0-alpha.14",{% endraw %}{% endif %}{% raw %}
     "@nuxtjs/eslint-config-typescript": "^12.1.0",
+    "@playwright/test": "^1.52.0",
     "@vitest/coverage-istanbul": "^3.1.3",{% endraw %}{% if frontend_uses_graphql %}{% raw %}
     "@vue/apollo-composable": "^4.2.2",{% endraw %}{% endif %}{% raw %}
     "@vue/devtools-api": "^7.7.2",
@@ -57,6 +57,7 @@
     "playwright-core": "^1.52.0",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.0.14",
+    "typescript": "{% endraw %}{{ typescript_version }}{% raw %}",
     "vitest": "^3.1.3",
     "vue-eslint-parser": "^10.1.1",
     "vue-tsc": "^2.2.8"
