+++ template/frontend/tests/e2e/helpers/playwright.ts
@@ -1,0 +1,8 @@
+import { inject } from "vitest";
+
+export function url(path: string): string {
+  if (!path.startsWith("/")) {
+    path = `/${path}`;
+  }
+  return `${inject("baseUrl")}${path}`;
+}
